requirements.txt
numpy>=1.6.1
scipy>=0.10.1
biopython>=1.61